* vc2005 scriop* warning fixes * lua is now shutdown on exit * warning fixes in scripts. * vc2005 script x64 configurations (debug+release, scripts will now function under x64.)
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2230 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/scripts/src/InstanceScripts/Raid_BlackwingLair.cpp
@@ -65,7 +65,7 @@ public:
 
     void AIUpdate()
     {
-		float val = sRand.rand(100.0f);
+		float val = (float)sRand.rand(100.0f);
         SpellCast(val);
     }
 
@@ -176,7 +176,7 @@ public:
 
     void AIUpdate()
     {
-		float val = sRand.rand(100.0f);
+		float val = (float)sRand.rand(100.0f);
         SpellCast(val);
     }
 
@@ -294,7 +294,7 @@ public:
 
     void AIUpdate()
     {
-		float val = sRand.rand(100.0f);
+		float val = (float)sRand.rand(100.0f);
         SpellCast(val);
     }
 
@@ -406,7 +406,7 @@ public:
 
     void AIUpdate()
     {
-		float val = sRand.rand(100.0f);
+		float val = (float)sRand.rand(100.0f);
         SpellCast(val);
     }
 
@@ -514,7 +514,7 @@ public:
 
     void AIUpdate()
     {
-		float val = sRand.rand(100.0f);
+		float val = (float)sRand.rand(100.0f);
         SpellCast(val);
     }
 
@@ -653,7 +653,7 @@ public:
 
     void AIUpdate()
     {
-		float val = sRand.rand(100.0f);
+		float val = (float)sRand.rand(100.0f);
         SpellCast(val);
     }
 
@@ -759,7 +759,7 @@ public:
 
     void AIUpdate()
     {
-		float val = sRand.rand(100.0f);
+		float val = (float)sRand.rand(100.0f);
         SpellCast(val);
     }
 
@@ -864,7 +864,7 @@ public:
 
     void AIUpdate()
     {
-		float val = sRand.rand(100.0f);
+		float val = (float)sRand.rand(100.0f);
         SpellCast(val);
     }
 
@@ -969,7 +969,7 @@ public:
 
     void AIUpdate()
     {
-		float val = sRand.rand(100.0f);
+		float val = (float)sRand.rand(100.0f);
         SpellCast(val);
     }
 
@@ -1123,7 +1123,7 @@ public:
 		}
 
 
-		float val = sRand.rand(100.0f);
+		float val = (float)sRand.rand(100.0f);
         SpellCast(val);
     }
